Add badges to developer profile pages
