+++ app/components/developers/card_component.rb
@@ -3,7 +3,6 @@ module Developers
     with_collection_parameter :developer
 
     delegate :featured?, to: :developer
-    delegate :recently_active?, to: :developer
 
     private attr_reader :developer, :highlight_featured
