+++ app/views/developers/show.html.erb
@@ -13,31 +13,31 @@
     <div class="pt-6 pb-8 xl:pb-10">
       <div class="max-w-3xl mx-auto px-4 sm:px-6 lg:px-8 lg:max-w-5xl lg:grid lg:grid-cols-3">
         <div class="lg:col-span-2 lg:pr-8 lg:border-r lg:border-gray-200">
-          <div>
-            <div>
-              <div class="md:flex md:items-start md:justify-between md:space-x-4 lg:border-b lg:pb-6">
-                <div class="w-full">
-                  <h1 class="text-2xl font-bold text-gray-900"><%= @developer.hero %></h1>
-                  <%= render Users::PaywalledComponent.new(user: current_user, paywalled: @developer, size: :small) do %>
-                    <div class="mt-2 flex items-center text-sm font-medium text-gray-500">
-                      <%= inline_svg_tag "icons/solid/user.svg", class: "flex-shrink-0 mr-1.5 h-5 w-5 text-gray-400", aria_hidden: true %>
-                      <%= @developer.name %>
-                    </div>
-                  <% end %>
-                </div>
-
-                <%= render Developers::PrimaryActionComponent.new(user: current_user, developer: @developer, business: current_user&.business) %>
+          <div class="space-y-6">
+            <div class="md:flex md:items-start md:justify-between md:space-x-4 lg:border-b lg:pb-6">
+              <div class="w-full">
+                <h1 class="text-2xl font-bold text-gray-900"><%= @developer.hero %></h1>
+                <%= render Users::PaywalledComponent.new(user: current_user, paywalled: @developer, size: :small) do %>
+                  <div class="mt-2 flex items-center text-sm font-medium text-gray-500">
+                    <%= inline_svg_tag "icons/solid/user.svg", class: "flex-shrink-0 mr-1.5 h-5 w-5 text-gray-400", aria_hidden: true %>
+                    <%= @developer.name %>
+                  </div>
+                <% end %>
               </div>
 
-              <aside class="mt-8 lg:hidden">
-                <%= render "details", developer: @developer %>
-              </aside>
+              <%= render Developers::PrimaryActionComponent.new(user: current_user, developer: @developer, business: current_user&.business) %>
+            </div>
 
-              <div class="py-3 lg:pt-6 lg:pb-0">
-                <h2 class="sr-only"><%= t(".description") %></h2>
-                <div class="prose max-w-prose">
-                  <%== @developer.rich_text_bio %>
-                </div>
+            <%= render Developers::BadgesComponent.new(@developer) %>
+
+            <aside class="mt-8 lg:hidden">
+              <%= render "details", developer: @developer %>
+            </aside>
+
+            <div class="pb-3 lg:pb-0">
+              <h2 class="sr-only"><%= t(".description") %></h2>
+              <div class="prose max-w-prose">
+                <%== @developer.rich_text_bio %>
               </div>
             </div>
           </div>
